# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  internal_api_r18devdump.searchResponse:
    properties:
      content_id:
        type: string
      dvd_id:
        type: string
      query:
        type: string
    type: object
...
host: localhost:8765
info:
  contact:
    name: API Support
    url: https://github.com/javinizer/javinizer-go/issues
  description: REST API for JAV metadata scraping and file organization
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/javinizer/javinizer-go
  title: Javinizer API
  version: "1.0"
paths:
...
  /api/v1/r18dev/dump:
    delete:
      description: Delete the local r18.dev dump sidecar database. After clearing,
        the scraper falls back to HTTP content_id resolution until the dump is re-downloaded.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Clear the r18.dev dump
      tags:
      - r18dev
  /api/v1/r18dev/dump/download:
    post:
      consumes:
      - application/json
      description: Start a full download of the r18.dev database dump and build the
        local lookup database. Progress is streamed via WebSocket. Returns 409 if
        a download is already running.
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Download the r18.dev dump
      tags:
      - r18dev
  /api/v1/r18dev/dump/search:
    get:
      description: Look up a dvd_id or content_id in the local r18.dev dump. Tries
        dvd_id first, then content_id.
      parameters:
      - description: dvd_id or content_id to look up
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_r18devdump.searchResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search the r18.dev dump
      tags:
      - r18dev
...
  /api/v1/r18dev/dump/update:
    post:
      consumes:
      - application/json
      description: Re-download the dump only if a newer version is available. Progress
        is streamed via WebSocket. Returns 409 if a download is already running.
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update the r18.dev dump
      tags:
      - r18dev
...
schemes:
- http
- https
swagger: "2.0"